refine the code acoriding review advice test=develop
