+++ python/paddle/fluid/tests/unittests/test_ones_op.py
@@ -22,58 +22,61 @@ import paddle
 import paddle.fluid.core as core
 from paddle.fluid.op import Operator
 import paddle.fluid as fluid
-from paddle.fluid import compiler, Program, program_guard
+import numpy as np
 
 
-class TestZerosOpError(unittest.TestCase):
-    def test_errors(self):
-        with program_guard(Program(), Program()):
-            # The input dtype of zeros_op must be bool, float16, float32, float64, int32, int64.
-            shape = [4]
-            dtype = "int8"
-            self.assertRaises(TypeError, fluid.layers.zeros, shape, dtype)
-
-
-class ApiZerosTest(unittest.TestCase):
+class ApiOnesTest(unittest.TestCase):
     def test_out(self):
         with fluid.program_guard(fluid.Program()):
-            zeros = paddle.zeros(shape=[10], dtype="float64")
+            ones = paddle.ones(shape=[10], dtype="float64")
             place = fluid.CPUPlace()
             exe = fluid.Executor(place)
-            result, = exe.run(fetch_list=[zeros])
-            expected_result = np.zeros(10, dtype="float64")
+            result, = exe.run(fetch_list=[ones])
+            expected_result = np.ones(10, dtype="float64")
         self.assertEqual((result == expected_result).all(), True)
 
         with fluid.program_guard(fluid.Program()):
-            zeros = paddle.zeros(shape=[10], dtype="int64")
+            ones = paddle.ones(shape=[10], dtype="int64")
             place = fluid.CPUPlace()
             exe = fluid.Executor(place)
-            result, = exe.run(fetch_list=[zeros])
-            expected_result = np.zeros(10, dtype="int64")
+            result, = exe.run(fetch_list=[ones])
+            expected_result = np.ones(10, dtype="int64")
         self.assertEqual((result == expected_result).all(), True)
 
         with fluid.program_guard(fluid.Program()):
-            zeros = paddle.zeros(shape=[10], dtype="int64")
+            ones = paddle.ones(shape=[10], dtype="int64", device="cpu")
             place = fluid.CPUPlace()
             exe = fluid.Executor(place)
-            result, = exe.run(fetch_list=[zeros])
-            expected_result = np.zeros(10, dtype="int64")
+            result, = exe.run(fetch_list=[ones])
+            expected_result = np.ones(10, dtype="int64")
         self.assertEqual((result == expected_result).all(), True)
 
 
-class ApiZerosError(unittest.TestCase):
+class ApiOnesZerosError(unittest.TestCase):
     def test_errors(self):
         def test_error1():
             with fluid.program_guard(fluid.Program()):
-                ones = fluid.layers.zeros(shape=10, dtype="int64")
+                ones = paddle.ones(shape=10, dtype="int64", device="opu")
 
-        self.assertRaises(TypeError, test_error1)
+        self.assertRaises(ValueError, test_error1)
 
         def test_error2():
             with fluid.program_guard(fluid.Program()):
-                ones = fluid.layers.zeros(shape=[10], dtype="int8")
+                ones = paddle.ones(shape=10, dtype="int64", device="opu")
 
-        self.assertRaises(TypeError, test_error2)
+        self.assertRaises(ValueError, test_error2)
+
+        def test_error3():
+            with fluid.program_guard(fluid.Program()):
+                ones = fluid.layers.ones(shape=10, dtype="int64")
+
+        self.assertRaises(TypeError, test_error3)
+
+        def test_error4():
+            with fluid.program_guard(fluid.Program()):
+                ones = fluid.layers.ones(shape=[10], dtype="int8")
+
+        self.assertRaises(TypeError, test_error4)
 
 
 if __name__ == "__main__":
